backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.7.1
python-multipart==0.0.9
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
